requirements.txt
discord.py==2.3.2
aiohttp
beautifulsoup4
flask
